code/requirements.txt
python-dotenv
flask 
langchain 
langchain-groq
langchain-community
langchain-postgres
psycopg[binary,pool]